requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
shortuuid
pydantic[dotenv]
qrcode[pil]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart